carWash/requirements.txt
Django==3.0.8
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
phonenumbers==8.12.6
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
django-requestlogs
django-ipware
